refactor(presentation): extract *DetailViewModelFactory classes into their own files
Move ViewModelFactory classes out of their corresponding ViewModel files for creature, magicalitem, and spell.
