+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/creature/presentation/viewmodel/CreatureDetailViewModel.kt
@@ -1,9 +1,7 @@
 package com.cyrillrx.rpg.creature.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.CreationExtras
 import com.cyrillrx.rpg.core.presentation.state.DetailState
 import com.cyrillrx.rpg.creature.domain.Creature
 import com.cyrillrx.rpg.creature.domain.CreatureRepository
@@ -11,7 +9,6 @@ import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.stateIn
-import kotlin.reflect.KClass
 
 class CreatureDetailViewModel(
     creatureId: String,
@@ -21,14 +18,4 @@ class CreatureDetailViewModel(
         val item = repository.getById(creatureId)
         emit(DetailState.of(creatureId, item))
     }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(stopTimeoutMillis = 5000), DetailState.Loading)
-}
-
-class CreatureDetailViewModelFactory(
-    private val creatureId: String,
-    private val repository: CreatureRepository,
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: KClass<T>, extras: CreationExtras): T {
-        @Suppress("UNCHECKED_CAST")
-        return CreatureDetailViewModel(creatureId, repository) as T
-    }
 }
