+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/magicalitem/presentation/viewmodel/MagicalItemDetailViewModel.kt
@@ -1,9 +1,7 @@
 package com.cyrillrx.rpg.magicalitem.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.CreationExtras
 import com.cyrillrx.rpg.core.presentation.state.DetailState
 import com.cyrillrx.rpg.magicalitem.domain.MagicalItem
 import com.cyrillrx.rpg.magicalitem.domain.MagicalItemRepository
@@ -11,7 +9,6 @@ import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.stateIn
-import kotlin.reflect.KClass
 
 class MagicalItemDetailViewModel(
     magicalItemId: String,
@@ -21,14 +18,4 @@ class MagicalItemDetailViewModel(
         val item = repository.getById(magicalItemId)
         emit(DetailState.of(magicalItemId, item))
     }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(stopTimeoutMillis = 5000), DetailState.Loading)
-}
-
-class MagicalItemDetailViewModelFactory(
-    private val magicalItemId: String,
-    private val repository: MagicalItemRepository,
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: KClass<T>, extras: CreationExtras): T {
-        @Suppress("UNCHECKED_CAST")
-        return MagicalItemDetailViewModel(magicalItemId, repository) as T
-    }
 }
